+++ CHANGES.rst
@@ -4,7 +4,12 @@ Changelog
 1.2 (unreleased)
 ----------------
 
-- No changes yet.
+- Set ``persist`` to ``False`` on all subwidgets created by
+  ``cron_edit_renderer`` for correct auto persistence.
+  [rnix]
+
+- No need to call cron widget when creating widget tree for proper extraction.
+  [rnix]
 
 
 1.1 (2018-07-16)